Move Protostream UTF test
