+++ protostream/src/main/java/org/infinispan/protostream/utf/StringWriter.java
@@ -1,4 +1,4 @@
-package org.infinispan;
+package org.infinispan.protostream.utf;
 
 import java.io.IOException;
